Move theme switch back to page
