+++ vizro-core/src/vizro/models/_components/graph.py
@@ -1,11 +1,10 @@
 import logging
 from typing import List, Literal
 
-from dash import Input, Output, Patch, callback, dcc
+from dash import dcc
 from plotly import graph_objects as go
 from pydantic import Field, PrivateAttr, validator
 
-import vizro._themes as themes
 import vizro.plotly.express as px
 from vizro.managers import data_manager
 from vizro.models import Action, VizroBaseModel
@@ -75,7 +74,6 @@ class Graph(VizroBaseModel):
 
     @_log_call
     def build(self):
-        self._update_graph_theme()
         self._set_callable_component()
         return dcc.Loading(
             dcc.Graph(
@@ -95,17 +93,6 @@ class Graph(VizroBaseModel):
             color="grey",
             parent_className="chart_container",
         )
-
-    def _update_graph_theme(self):
-        @callback(
-            Output(self.id, "figure", allow_duplicate=True),
-            Input("theme_selector", "on"),
-            prevent_initial_call="initial_duplicate",
-        )
-        def update_graph_theme(theme_selector_on: bool):
-            patched_figure = Patch()
-            patched_figure["layout"]["template"] = themes.dark if theme_selector_on else themes.light
-            return patched_figure
 
     def _set_callable_component(self):
         self._callable_component = self.figure
